clipboard-apis/clipboardevent_interface_synthetic_paste_event_does_not_insert_plain_text_data_into_contenteditable_element_paste_on_document.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - ClipboardEvent interface - synthetic paste event does not insert plain text data into contentEditable element   - paste on document</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** ClipboardEvent interface - synthetic paste event does not insert plain text data into contentEditable element   */
/* Events: paste */
/* paste data: "clipboard text" */
/* Test HTML: <div id="editable" contentEditable><p>Editable element</p></div><script>document.getElementById('editable').focus();<\/script>*/
/* Targets: document */
function clipboard_api_test(e){}
function onTestSetupReady(type, test_obj){
    var evt = new ClipboardEvent(type);
    evt.clipboardData.items.add('synthetic event text', 'text/plain');
    document.getElementById('editable').dispatchEvent(evt);
    test_obj.step_timeout( function(){
      result(function(){
        assert_equals(document.getElementById('editable').textContent.indexOf( 'synthetic event text' ), -1);
      });
    }, 50 );
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <div id="editable" contentEditable><p>Editable element</p></div><script>document.getElementById('editable').focus();</script>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste="clipboard text" ;
        var extPassCond=null;
        var eventTarget=document;
        var eventType='paste';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
